backend/requirements/dev.txt
-r base.txt

django-debug-toolbar==5.2.0
django-extensions==4.1
ipython==9.4.0
